% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{construct_user_data_path}
\alias{construct_user_data_path}
\title{Construct user app data directory path}
\usage{
construct_user_data_path()
}
\value{
Character. Path to user's app data directory
}
\description{
Consists of three pieces of information:

- System location
- App name
- App version
}
