% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_metadata.R
\name{convert_exprs_to_stata}
\alias{convert_exprs_to_stata}
\title{Convert expressions in target column from SuSo C# to Stata}
\usage{
convert_exprs_to_stata(suso_expr, varname)
}
\arguments{
\item{suso_expr}{Character vector. SuSo enablement or validation condition.}

\item{varname}{Character vector. Variable name to replace `self`}
}
\value{
Data frame. Questionnaire metadata with translated expressions.
}
\description{
Convert expressions in target column from SuSo C# to Stata
}
\details{
Apply series of functions to translate C# to equivalent Stata
}
