% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_metadata.R
\name{create_detailed_q_type}
\alias{create_detailed_q_type}
\title{Create detailed question type}
\usage{
create_detailed_q_type(
  qnr_df,
  type_var = "type",
  mask_var = "mask",
  is_integer_var = "is_integer",
  linked_to_roster_id_var = "linked_to_roster_id",
  linked_to_question_id_var = "linked_to_question_id",
  yes_no_view_var = "yes_no_view",
  is_timestamp_var = "is_timestamp"
)
}
\arguments{
\item{qnr_df}{Data frame. Questionnaire metadata.}

\item{type_var}{Character. Name of column with data type from JSON.}

\item{mask_var}{Character. Name of column indicating whether text is masked.}

\item{is_integer_var}{Character. Name of column whether is integer.}

\item{linked_to_roster_id_var}{Name of column with linked roster ID.}

\item{linked_to_question_id_var}{Name of column with linked question.}

\item{yes_no_view_var}{Name of column with yes/no mode indicator.}

\item{is_timestamp_var}{Name of column saying whether date is timestamp.}
}
\value{
Data frame. Questionnaire metadata with `type_detailed` column.
}
\description{
Create detailed question type
}
