% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_user_app_dir}
\alias{create_user_app_dir}
\title{Create user application directory for persistent storage}
\usage{
create_user_app_dir()
}
\value{
Character. Path where app directory was created/exists
}
\description{
Create user application directory for persistent storage
}
