% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_prepare_for_vars.R
\name{prepare_for_vars}
\alias{prepare_for_vars}
\title{Prepare the data for variables of interest}
\usage{
prepare_for_vars(
  qnr_df,
  vars,
  rc_expr_target = "@rowcode",
  rc_expr_replacement = "some__id",
  rt_lbl_target = "\%rostertitle\%",
  rt_lbl_replacement = "[ITEM]",
  oth_var_pattern,
  show_vars = "interview__id"
)
}
\arguments{
\item{qnr_df}{Data frame. Variable metadata.}

\item{vars}{Character vector. Names of variables.}

\item{rc_expr_target}{Character. `rowcode` variable in expressoins (e.g., `"@rowcode"`).}

\item{rc_expr_replacement}{Character. Replacement for rowcode.}

\item{rt_lbl_target}{Character. Pipe to replace in variable label.}

\item{rt_lbl_replacement}{Character. Replacement for roster title.}

\item{oth_var_pattern}{Character. Regex for identifying 
"other (specify)" variables}

\item{show_vars}{Character. Variables to show as context.}
}
\value{
Data frame. Questionnaire metadata prepared for grou.
}
\description{
Prepare the data for variables of interest
}
