% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_metadata.R
\name{prepare_metadata}
\alias{prepare_metadata}
\title{Prepare metadata}
\usage{
prepare_metadata(qnr_df)
}
\arguments{
\item{qnr_df}{Data frame. Questionnaire metadata.}
}
\value{
Data frame. Modified questinnaire metadata file.
}
\description{
Prepare metadata
}
