% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_metadata.R
\name{sanitize_text}
\alias{sanitize_text}
\title{Sanitize variable label}
\usage{
sanitize_text(qnr_df)
}
\arguments{
\item{qnr_df}{Data frame. Questionnaire metadata.}
}
\value{
Data frame. Questionnaire metadata with sanitized `label` column.
}
\description{
Remove HTML tags
}
