% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_write_var_block.R
\name{write_answer_options}
\alias{write_answer_options}
\title{Write answer options in template}
\usage{
write_answer_options(var_df)
}
\arguments{
\item{var_df}{Data frame. Contains single row for variable of interest.}
}
\value{
Character. Answer option block. That looks like this:
ANSWERS:
1    Answer 1
2    Answer 2
N    Answer N
}
\description{
Write answer options in template
}
