% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_write_var_block.R
\name{write_extreme_val_txt}
\alias{write_extreme_val_txt}
\title{Write extreme values code block}
\usage{
write_extreme_val_txt(type, varname, if_condition, show_vars)
}
\arguments{
\item{type}{Character. Variable type string created by 
`create_detailed_q_type()`.}

\item{varname}{Character. Variable name.}

\item{if_condition}{Character. Stata if condition.}

\item{show_vars}{Character. Variables to show for context.}
}
\value{
Character string. Either filled template or empty string, on
variable type in `type`.
}
\description{
Write extreme values code block
}
