% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_write_var_block.R
\name{write_if_condition}
\alias{write_if_condition}
\title{Write if-condition code component}
\usage{
write_if_condition(condition_expression)
}
\arguments{
\item{condition_expression}{Character. Stata condition.}
}
\value{
Character string. Either filled template or empty string, on
whether `condition_expression` is NA/empty or not.
}
\description{
Write if-condition code component
}
