% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_write_var_block.R
\name{write_oth_val_txt}
\alias{write_oth_val_txt}
\title{Write other values code block}
\usage{
write_oth_val_txt(varname, type, oth_var_pattern, if_condition)
}
\arguments{
\item{varname}{Character. Variable name.}

\item{type}{Character. Detailed variable type.}

\item{oth_var_pattern}{Character. (Regex) expression to identify 
"other (specify)" variables.}

\item{if_condition}{Character. Stata if condition.}
}
\value{
Character string. Either filled template or empty string, on
whether variable name matches the pattern.
}
\description{
Write other values code block
}
