% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_write_var_block.R
\name{write_validation_txt}
\alias{write_validation_txt}
\title{Write validation code block}
\usage{
write_validation_txt(var_df)
}
\arguments{
\item{var_df}{Data frame. Questionnaire metadata filtered to single 
variable of interest.}
}
\value{
Character. String with templated validaiton condition. 
Either empty or filled.
}
\description{
Write validation code block
}
