% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_write_var_block.R
\name{write_var_block}
\alias{write_var_block}
\title{Write code block for a variable}
\usage{
write_var_block(qnr_df, var)
}
\arguments{
\item{qnr_df}{Data frame. Modified questionnaire metadata for 
a group of variables.}

\item{var}{Character. Name of a variable whose code block to write.}
}
\value{
Character. Code block for a variable
}
\description{
Write code block for a variable
}
